services/tasks/pyproject.toml
[tool.poetry]
name = "tasks-service"
version = "0.1.0"
description = "Tasks microservice"
authors = ["Your Name <your@email.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
pydantic = "^2.6.0"
pydantic-settings = "^2.2.1"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
asyncpg = "^0.29.0"
SQLAlchemy = "^2.0.29"
alembic = "^1.13.1"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"
psycopg2-binary = "^2.9.9"

[tool.poetry.dev-dependencies]
pytest = "^8.1.1"
pytest-asyncio = "^0.23.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
